# client/src/main/resources/api-docs-standalone-service.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "Standalone REST API",
    "description" : "Standalone API specification.",
    "version" : "0.0.1"
  },
  "servers" : [ {
    "url" : "http://localhost:57732/standalone-microservice",
    "description" : "Generated server url"
  } ],
  "tags" : [ {
    "name" : "Standalone",
    "description" : "Standalone endpoint"
  } ],
  "paths" : {
    "/api/v1/properties" : {
      "get" : {
        "tags" : [ "Standalone" ],
        "summary" : "Returns list of properties.",
        "operationId" : "getProperties",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/StandaloneModel1"
                  }
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/StandaloneModel1"
                  }
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Standalone" ],
        "summary" : "Create a property",
        "operationId" : "addProperty",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/StandaloneModel3"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1" : {
      "get" : {
        "tags" : [ "Standalone" ],
        "summary" : "Get the microservice name",
        "operationId" : "getMicroserviceName",
        "parameters" : [ {
          "name" : "X-Client-Id",
          "in" : "header",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "X-Client-Ip",
          "in" : "header",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/properties/{id}" : {
      "delete" : {
        "tags" : [ "Standalone" ],
        "summary" : "Create a property",
        "operationId" : "deleteProperty",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "Property has been deleted"
          },
          "404" : {
            "description" : "Property does not exist"
          }
        }
      }
    }
  },
  "components" : {
    "schemas" : {
      "StandaloneModel3" : {
        "type" : "object",
        "properties" : {
          "standalone3Property1" : {
            "type" : "string"
          },
          "standalone3Property2" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      },
      "StandaloneModel1" : {
        "required" : [ "standaloneProperty2", "standalonePropertyId" ],
        "type" : "object",
        "properties" : {
          "standalonePropertyId" : {
            "type" : "string"
          },
          "standaloneProperty1" : {
            "maximum" : 100,
            "minimum" : 0,
            "type" : "integer",
            "format" : "int64"
          },
          "standaloneProperty2" : {
            "maxLength" : 20,
            "minLength" : 1,
            "type" : "string"
          },
          "standaloneList" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/StandaloneModel2"
            }
          }
        }
      },
      "StandaloneModel2" : {
        "type" : "object",
        "properties" : {
          "standalone2Property1" : {
            "type" : "string"
          },
          "standalone2Property2" : {
            "type" : "integer",
            "format" : "int32"
          }
        }
      }
    }
  }
}